Tiny bit of cleanup

+++ p_local.h
@@ -255,7 +255,7 @@ extern	int			bmapwidth, bmapheight;	/* in mapblocks */
 extern	fixed_t		bmaporgx, bmaporgy;		/* origin of block map */
 extern	mobj_t		**blocklinks;			/* for thing chains */
 
-extern	int			numthings;
+extern	uint16_t			numthings;
 extern	spawnthing_t* spawnthings;
 
 extern	VINT		*validcount;		/* (0 - increment every time a check is made, [1..numlines]) x 2 */
